api_test2/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
djangorestframework-jwt